refactor(doctor): split doctor.rs into many files
