+++ src/doctor/mod.rs
@@ -1,0 +1,7 @@
+pub mod inspect;
+pub mod path;
+pub mod prerequisites;
+pub mod rules;
+pub mod services;
+
+pub use crate::doctor::inspect::{inspect, run_cli, DiagnosticStatus};
